TDD/vending-machine/src/main/java/com/spark/vending/product/ProductService.java
package com.spark.vending.product;

import java.util.List;

import org.springframework.stereotype.Service;

@Service
public class ProductService {
	
	private ProductRepository productRepository;
	
	public ProductService(ProductRepository productRepository) {
		this.productRepository = productRepository;
	}
	
	public Product addProduct(Product product) {
		return productRepository.save(product);
	}
	
	public List<Product> getProducts(){
		return productRepository.findAll();
	}
	
	public Product getProductByName(String name) {
		return productRepository.findProductByName(name);
	}
	
	
}
